+++ crawlib/entity/base.py
@@ -243,13 +243,6 @@ class EntityExtendScheduler(EntityBase):
         return get_all_subclass(cls)
 
     @classmethod
-    def validate_implementation_orm_related(cls):
-        """
-        Run ORM framework specified implementation validation.
-        """
-        pass
-
-    @classmethod
     def _validate_configuration(cls):
         """
         Check if the subclass of :class:`EntityOrm` is correctly implemented.
@@ -266,36 +259,8 @@ class EntityExtendScheduler(EntityBase):
         if not isinstance(cls.CONF_EDIT_AT_KEY, six.string_types):
             raise TypeError("`CONF_EDIT_AT_KEY` has to be a str!")
 
-        # msg = ("you have to specify `CONF_UPDATE_FIELDS`!"
-        #        "it is the tuple of field name that you want to update "
-        #        "when you crawled the url generated by cls.build_url(**kwargs)")
-        # if cls.CONF_UPDATE_FIELDS is None:
-        #     raise NotImplementedError(msg)
-        # elif isinstance(cls.CONF_UPDATE_FIELDS, (tuple, list)):
-        #     for field in cls.CONF_UPDATE_FIELDS:
-        #         if not hasattr(cls, field):
-        #             msg = ("`CONF_UPDATE_FIELDS` has '{}', but attribute '{}' "
-        #                    "not found in {}").format(
-        #                 field, field, cls
-        #             )
-        #             raise NotImplementedError(msg)
-        #
-        #     tmp_lst = list(cls.CONF_UPDATE_FIELDS)
-        #     if cls._ORM_FRAMEWORK == "mongoengine":
-        #         if "_id" not in tmp_lst:
-        #             tmp_lst.append("_id")
-        #     if cls.CONF_STATUS_KEY not in tmp_lst:
-        #         tmp_lst.append(cls.CONF_STATUS_KEY)
-        #     if cls.CONF_EDIT_AT_KEY not in tmp_lst:
-        #         tmp_lst.append(cls.CONF_EDIT_AT_KEY)
-        #     cls.CONF_UPDATE_FIELDS = tuple(tmp_lst)
-        # else:
-        #     raise ValueError(msg)
-
-        cls.validate_implementation_orm_related()
-
     @classmethod
-    def validate_relationship_config(cls):
+    def _validate_relationship_config(cls):
         """
         Validate ``cls.CONF_RELATIONSHIP``
 
@@ -642,6 +607,7 @@ class Entity(EntityExtendScheduler):
         cls._validate_abstract_methods()
         cls._validate_configuration()
         cls._validate_orm_related()
+        cls._validate_relationship_config()
 
 
 @attr.s
